Create a nested reducer for owners
